feat: Add BarChartNoPadding component for displaying code line counts
Add a new component called BarChartNoPadding to display the number of lines of code for different programming languages. This component uses the recharts library for creating a bar chart with responsive layout. The component also includes SCSS styles for the chart container and title. This refactor improves the visualization of code line counts and enhances the overall design of the application.